refactor(cleanroom): all-code-review指摘の機械的修正(冗長cast削除/死にCountConnectors削除/テストの#region Internal除去)

+++ moorestech_server/Assets/Scripts/Game.Block/Factory/BlockTemplate/VanillaCleanRoomMachineTemplate.cs
@@ -71,7 +71,7 @@ namespace Game.Block.Factory.BlockTemplate
                 machineParam.ModuleSlotCount, _blockInventoryUpdateEvent, blockInstanceId,
                 machineParam.InputSlotCount, machineParam.OutputSlotCount);
 
-            var requestPower = (float)machineParam.RequiredPower;
+            var requestPower = machineParam.RequiredPower;
 
             // 新規作成またはセーブから復元
             // Create new or restore from save
